REVERT: wip on completing new extraction class, need to test TM, make template results, refine template and prepare stack match templates (the latter two with both images and coord files) and then also peak upsampling needs to be debugged. DO NOT USE THIS COMMIT for work.

+++ src/core/core_headers.h
@@ -17,7 +17,7 @@ struct Peak {
     Peak( ) = default; // Peak () {}; would also work
 
     // We could skip both ctors but by declaring this one it (helps) to avoid a mixup in ordering
-    Peak(float x_, long y_, float z_, float value_, long physical_address_within_image_)
+    Peak(float x_, float y_, float z_, float value_, long physical_address_within_image_)
         : x(x_), y(y_), z(z_), value(value_), physical_address_within_image(physical_address_within_image_) {}
 };
